+++ src/app/(days)/day2/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Text } from "react-native";
+
+
+export default  function DayDetailsScreen(){
+    return(
+        <Text>
+                day detail screen
+        </Text>
+    )
+}
